+++ Application/__init__.py
@@ -2,7 +2,7 @@
 Application.__init__.py
 A skeleton for a modern webapp
 """
-from . import (admin, app, assets, auth, database, extensions, models,
+from . import (admin, app, assets, auth, database, models,
                settings, views, utils)
 
 create_app = app.create_app
