+++ Application/auth/login_manager.py
@@ -1,0 +1,7 @@
+"""
+login_manager.py
+defines login manager
+"""
+from flask_login import LoginManager
+
+login_manager = LoginManager()
